PPTP-445 Rename packages and tidy app.Routes file
