+++ app/uk/gov/hmrc/plasticpackagingtax/returns/views/partials/siteHeader.scala.html
@@ -17,7 +17,7 @@
 @import uk.gov.hmrc.hmrcfrontend.views.html.components.HmrcHeader
 @import uk.gov.hmrc.hmrcfrontend.views.html.components.Header
 @import uk.gov.hmrc.hmrcfrontend.views.viewmodels.language
-@import uk.gov.hmrc.plasticpackagingtax.returns.controllers.{routes => pptRoutes}
+@import uk.gov.hmrc.plasticpackagingtax.returns.controllers.home.{routes => homeRoutes}
 
 
 @this(hmrcHeader: HmrcHeader)
@@ -28,7 +28,7 @@
 @hmrcHeader(Header(
 homepageUrl = "https://www.gov.uk",
 serviceName = Some(messages("service.name")),
-serviceUrl = pptRoutes.HomeController.displayPage.url,
+serviceUrl = homeRoutes.HomeController.displayPage.url,
 language = language.En,
 containerClasses = "govuk-width-container"
 ))
