+++ app/uk/gov/hmrc/plasticpackagingtax/returns/views/returns/start_page.scala.html
@@ -17,12 +17,13 @@
 @import uk.gov.hmrc.govukfrontend.views.html.components.GovukInsetText
 @import uk.gov.hmrc.govukfrontend.views.html.components.GovukButton
 @import uk.gov.hmrc.plasticpackagingtax.returns.views.model.Title
-@import uk.gov.hmrc.plasticpackagingtax.returns.controllers.{routes => pptRoutes}
+@import uk.gov.hmrc.plasticpackagingtax.returns.controllers.returns.{routes => returnsRoutes}
 @import uk.gov.hmrc.plasticpackagingtax.returns.views.html.components.bulletList
 @import uk.gov.hmrc.plasticpackagingtax.returns.views.html.components.pageTitle
 @import uk.gov.hmrc.plasticpackagingtax.returns.views.html.components.link
 @import uk.gov.hmrc.plasticpackagingtax.returns.views.html.components.sectionHeader
 @import uk.gov.hmrc.plasticpackagingtax.returns.views.html.components.paragraphBody
+@import uk.gov.hmrc.plasticpackagingtax.returns.views.html.main_template
 
 @this(
 govukLayout: main_template,
@@ -84,7 +85,7 @@ govukButton: GovukButton
 
         @govukButton(
             Button(
-                href = Some(pptRoutes.StartController.displayPage().url),
+                href = Some(returnsRoutes.StartController.displayPage().url),
                 isStartButton = true,
                 content = Text(messages("startPage.buttonName"))
             )
