+++ conf/home.routes
@@ -1,0 +1,3 @@
+# microservice specific routes
+
+GET        /                    uk.gov.hmrc.plasticpackagingtax.returns.controllers.home.HomeController.displayPage()
